Step 3c (drift fix): move verdict/lifecycle types from model/ to verdict/; delete model/
Per DIR-PACKAGE-DRIFT-FIX-punit step 3c. Four types (and their tests)
move from org.javai.punit.model to org.javai.punit.verdict where they
sit alongside the existing verdict abstractions:

- TerminationReason (verdict-side: how a run ended)
- BaselineProvenance (verdict-side: which baseline informed the test)
- ExpirationStatus (lifecycle: baseline freshness state)
- ExpirationPolicy (lifecycle: baseline freshness policy)

(FailureObservation was the fifth member of this group per the
original directive; it was deleted in DIR-CONTRACT-PACKAGE-REMOVAL-punit
as zero-callers dead code, so this commit doesn't move it.)

After this step both model/ trees (main and test) are empty and
deleted.

Touched:
- git mv each main and test file from model/ to verdict/.
- Rewrite their `package` declarations.
- Bulk-rewrite import paths in every module (punit-core, punit-junit5,
  punit-report, punit-sentinel) and in punitexamples — ~52 files.
- Drop the dead `org.javai.punit.model..` entry from
  CoreArchitectureTest's three rules and the equivalent in
  ExampleArchitectureTest's `infrastructureMustNotDependOnTestFramework`.
  PackageStructureArchitectureTest's noTopLevelModelPackage rule
  stays intact as a permanent regression guard now that its
  freeze-store has emptied.
- Drop the model row from docs/DEVELOPER-API.md's package-visibility
  table (model/ no longer exists).

Architecture-test bookkeeping:
- noTopLevelModelPackage rule's freeze-store: empty — at target state.
- engineUnderInternal rule's store: unchanged (no movement into
  engine/).
- Other rules' stores: unchanged.

Tests green across all four punit modules; punitexamples
composite build green.

Closes ArchUnit rule: noTopLevelModelPackage.
