--- punit-core/archunit_store/90083594-5684-45be-a3cf-87dc0d2f9509
@@ -1,9 +1,0 @@
-Class <org.javai.punit.model.BaselineProvenance> does reside in a package 'org.javai.punit.model..' in (BaselineProvenance.java:0)
-Class <org.javai.punit.model.ExpirationPolicy> does reside in a package 'org.javai.punit.model..' in (ExpirationPolicy.java:0)
-Class <org.javai.punit.model.ExpirationStatus$Expired> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
-Class <org.javai.punit.model.ExpirationStatus$ExpiringImminently> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
-Class <org.javai.punit.model.ExpirationStatus$ExpiringSoon> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
-Class <org.javai.punit.model.ExpirationStatus$NoExpiration> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
-Class <org.javai.punit.model.ExpirationStatus$Valid> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
-Class <org.javai.punit.model.ExpirationStatus> does reside in a package 'org.javai.punit.model..' in (ExpirationStatus.java:0)
-Class <org.javai.punit.model.TerminationReason> does reside in a package 'org.javai.punit.model..' in (TerminationReason.java:0)
