+++ punit-core/src/main/java/org/javai/punit/controls/budget/CostBudgetMonitor.java
@@ -2,7 +2,7 @@ package org.javai.punit.controls.budget;
 
 import java.util.Optional;
 import org.javai.punit.api.BudgetExhaustedBehavior;
-import org.javai.punit.model.TerminationReason;
+import org.javai.punit.verdict.TerminationReason;
 
 /**
  * Monitors time and token budgets for a probabilistic test method.
