+++ punit-core/src/main/java/org/javai/punit/spec/expiration/ExpirationReportPublisher.java
@@ -3,8 +3,8 @@ package org.javai.punit.spec.expiration;
 import java.time.Instant;
 import java.util.LinkedHashMap;
 import java.util.Map;
-import org.javai.punit.model.ExpirationPolicy;
-import org.javai.punit.model.ExpirationStatus;
+import org.javai.punit.verdict.ExpirationPolicy;
+import org.javai.punit.verdict.ExpirationStatus;
 import org.javai.punit.spec.model.ExecutionSpecification;
 
 /**
