+++ punit-core/src/main/java/org/javai/punit/spec/expiration/ExpirationWarningRenderer.java
@@ -4,8 +4,8 @@ import java.time.Duration;
 import java.time.Instant;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
-import org.javai.punit.model.ExpirationPolicy;
-import org.javai.punit.model.ExpirationStatus;
+import org.javai.punit.verdict.ExpirationPolicy;
+import org.javai.punit.verdict.ExpirationStatus;
 import org.javai.punit.reporting.DurationFormat;
 import org.javai.punit.reporting.PUnitReporter;
 import org.javai.punit.spec.model.ExecutionSpecification;
