+++ punit-core/src/main/java/org/javai/punit/spec/expiration/WarningLevel.java
@@ -1,6 +1,6 @@
 package org.javai.punit.spec.expiration;
 
-import org.javai.punit.model.ExpirationStatus;
+import org.javai.punit.verdict.ExpirationStatus;
 import org.javai.punit.statistics.transparent.TransparentStatsConfig.DetailLevel;
 
 /**
