+++ punit-core/src/main/java/org/javai/punit/verdict/ProbabilisticTestVerdict.java
@@ -10,8 +10,8 @@ import java.util.Optional;
 import org.javai.punit.api.TestIntent;
 import org.javai.punit.api.spec.FailureCount;
 import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
-import org.javai.punit.model.ExpirationStatus;
-import org.javai.punit.model.TerminationReason;
+import org.javai.punit.verdict.ExpirationStatus;
+import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.api.UseCaseAttributes;
 
 /**
