+++ punit-core/src/main/java/org/javai/punit/verdict/ProbabilisticTestVerdictBuilder.java
@@ -17,8 +17,8 @@ import org.javai.punit.api.UseCaseAttributes;
 import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
 import org.javai.punit.controls.budget.SharedBudgetMonitor;
 import org.javai.punit.controls.pacing.PacingConfiguration;
-import org.javai.punit.model.ExpirationStatus;
-import org.javai.punit.model.TerminationReason;
+import org.javai.punit.verdict.ExpirationStatus;
+import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.reporting.RateFormat;
 import org.javai.punit.spec.expiration.ExpirationEvaluator;
 import org.javai.punit.spec.model.ExecutionSpecification;
