+++ punit-core/src/main/java/org/javai/punit/verdict/VerdictAdapter.java
@@ -11,7 +11,7 @@ import org.javai.punit.api.spec.ProbabilisticTestResult;
 import org.javai.punit.api.spec.Verdict;
 import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
 import org.javai.punit.engine.output.LatencySection;
-import org.javai.punit.model.TerminationReason;
+import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.verdict.ProbabilisticTestVerdictBuilder.LatencyInput;
 import org.javai.punit.verdict.ProbabilisticTestVerdictBuilder.MisalignmentInput;
