+++ punit-core/src/test/java/org/javai/punit/controls/budget/CostBudgetMonitorTest.java
@@ -3,7 +3,7 @@ package org.javai.punit.controls.budget;
 import static org.assertj.core.api.Assertions.assertThat;
 import java.util.Optional;
 import org.javai.punit.api.BudgetExhaustedBehavior;
-import org.javai.punit.model.TerminationReason;
+import org.javai.punit.verdict.TerminationReason;
 import org.junit.jupiter.api.Test;
 
 /**
