+++ punit-core/src/test/java/org/javai/punit/spec/expiration/ExpirationWarningRendererTest.java
@@ -3,8 +3,8 @@ package org.javai.punit.spec.expiration;
 import static org.assertj.core.api.Assertions.assertThat;
 import java.time.Duration;
 import java.time.Instant;
-import org.javai.punit.model.ExpirationPolicy;
-import org.javai.punit.model.ExpirationStatus;
+import org.javai.punit.verdict.ExpirationPolicy;
+import org.javai.punit.verdict.ExpirationStatus;
 import org.javai.punit.spec.model.ExecutionSpecification;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
