+++ punit-core/src/test/java/org/javai/punit/verdict/ProbabilisticTestVerdictBuilderTest.java
@@ -14,8 +14,8 @@ import org.javai.punit.controls.budget.SharedBudgetMonitor;
 import org.javai.punit.controls.budget.SharedBudgetMonitor.Scope;
 import org.javai.punit.api.BudgetExhaustedBehavior;
 import org.javai.punit.controls.pacing.PacingConfiguration;
-import org.javai.punit.model.ExpirationPolicy;
-import org.javai.punit.model.TerminationReason;
+import org.javai.punit.verdict.ExpirationPolicy;
+import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.spec.model.ExecutionSpecification;
 import org.javai.punit.statistics.BinomialProportionEstimator;
 import org.javai.punit.statistics.transparent.BaselineData;
