+++ punit-report/src/test/java/org/javai/punit/report/VerdictXmlWriterTest.java
@@ -25,8 +25,8 @@ import org.javai.punit.api.TestIntent;
 import org.javai.punit.api.spec.FailureCount;
 import org.javai.punit.api.spec.FailureExemplar;
 import org.javai.punit.controls.budget.CostBudgetMonitor.TokenMode;
-import org.javai.punit.model.ExpirationStatus;
-import org.javai.punit.model.TerminationReason;
+import org.javai.punit.verdict.ExpirationStatus;
+import org.javai.punit.verdict.TerminationReason;
 import org.javai.punit.api.UseCaseAttributes;
 import org.javai.punit.verdict.ProbabilisticTestVerdict;
 import org.javai.punit.verdict.ProbabilisticTestVerdict.*;
